codereview: moved cli validation to separate file
